% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling.R
\name{summarize_rolling}
\alias{summarize_rolling}
\title{Title}
\usage{
summarize_rolling(
  .data,
  .interval,
  na.rm = TRUE,
  datevar,
  groupvar,
  summvar,
  summfunction,
  ...
)
}
\arguments{
\item{.data}{The data passed through dplyr}

\item{.interval}{An interval generated by interval_rolling}

\item{na.rm}{Remove data in summarize?}

\item{datevar}{The variable in .data that stores date information}

\item{groupvar}{The variable in .data that is used for grouping}

\item{summvar}{The variable in .data that is used for summarization}

\item{summfunction}{The function to apply to summvar}

\item{...}{}
}
\value{
Data
}
\description{
Title
}
\examples{
NA
}
